requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.6.1
python-jose==3.3.0
requests==2.32.3
python-dotenv==1.0.1
